requirements.txt
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.21
openai>=1.3.0
python-dotenv>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.1
pypdf>=3.16.4
docx2txt>=0.8
tiktoken>=0.5.1
pandas>=2.1.1
faiss-cpu>=1.7.4
tenacity>=8.2.3
pytest>=7.4.2
httpx>=0.25.0
beautifulsoup4>=4.12.2
requests>=2.31.0
psutil>=5.9.0
crawl4ai>=0.1.0
